backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.3.4
qdrant-client==1.9.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
slowapi==0.1.8